Refactor execution of tasks to use multiple transactions. Refine readme.
