+++ src/main/resources/application.properties
@@ -2,10 +2,8 @@ server.port=8090
 spring.datasource.url=jdbc:h2:~/testautomationservice
 spring.datasource.username=root
 spring.datasource.password=
-
 spring.datasource.connectionProperties=useUnicode=true;characterEncoding=utf-8;
 spring.jpa.hibernate.ddl-auto=update
 spring.h2.console.enabled=true
 spring.h2.console.path=/console/
-
 spring.activemq.broker-url=tcp://localhost:61616
